Move allocator asignement into macro.
